user-settings-screen - Added UserSettingsScreen.
