+++ lib/ui/screens/intro_screen.dart
@@ -1,5 +1,5 @@
 import 'package:cookie/ui/screens/auth_screen.dart';
-import 'package:cookie/ui/screens/profile_screen.dart';
+import 'package:cookie/ui/screens/user_settings_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -27,7 +27,7 @@ class _IntroScreenState extends State<IntroScreen> {
       context,
       MaterialPageRoute(
         builder:
-            (context) => ProfileScreen(uid: result.uid),),)
+            (context) => UserSettingsScreen(uid: result.uid),),)
 
         : Navigator.pushReplacement(
       context,
