+++ pubspec.yaml
@@ -60,6 +60,7 @@ flutter:
     - assets/images/cookie.png
     - assets/images/blurred_cookie.png
     - assets/images/background_profile_screen.png
+    - assets/images/little_cookie.png
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
